requirements.txt
bottle<1
requests
schedule
PyMySQL3
